+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Removed the `include` argument from the `addLogo()` and `addPhoto()` methods
 - Removed unnecessary comments
 - Added the missing functionality to the `addLogo()` and `addPhoto()` methods
+- Added the missing functionality to the `addLogoURL()` and `addPhotoURL()` methods
 
 ## 0.3.4